+++ Fabulous.XamarinForms/samples/FabulousWeather/iOS/AppDelegate.fs
@@ -12,7 +12,6 @@ type AppDelegate () =
 
     override this.FinishedLaunching (app, options) =
         UIApplication.SharedApplication.SetStatusBarStyle(UIStatusBarStyle.LightContent, true)
-        Forms.SetFlags("IndicatorView_Experimental")
         Forms.Init()
         this.LoadApplication (new FabulousWeather.App())
         base.FinishedLaunching(app, options)