+++ Fabulous.XamarinForms/samples/LoginShape/LoginShape/LoginShape.fs
@@ -154,8 +154,6 @@ module App =
 type App () as app =
     inherit Application ()
 
-    do Device.SetFlags([| "Shapes_Experimental"; "AppTheme_Experimental" |]);
-
     let runner =
         App.program
 #if DEBUG